career_mate_backend/requirements.txt
annotated-types==0.7.0
bcrypt==4.0.1
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Mail==0.9.1
Flask-Session==0.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.4.0
pyparsing==3.2.5
PyPDF2==3.0.0
python-dotenv==1.0.0
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
httpx==0.27.0
Flask-Limiter==3.8.0
PyGithub==2.8.1
requests-oauthlib==2.0.0